[add] dynamic channel handling per connection process
when starting a connection worker we are going to start within it a pool
of channel workers in which each one is going to start a multiplexed
channel to RabbitMQ.
Each channel worker is going to be linked to a channel pid, every
connection worker is going to be linked to each channel worker
but trapping exits. this way if a channel crashes, the connection worker
is going to be able to start another channel and if a connection
to RabbitMQ crashes we are going to be able to restart that connection,
remove all crashed channels and then restarting them with a new connection.

[add] DynamicSupervisor for handling RabbitMQ channel pooling [add] start channels (default 1000) when stablishin connection to RabbitMQ
[add] each connection worker handles a pool of channels to serve