[refactor] remove dynamic supervisor for channels
now when starting a connection worker we are going to start within it a pool
of multiplexed channels to RabbitMQ and store them in its state (we can move
this later to ets). Then, inside the connection worker we are going to link
each channel but it's going to be trapping exits. this way if a channel crashes,
the connection worker is going to be able to start another channel and if a
connection to RabbitMQ crashes we are going to be able to restart that connection,
remove all crashed channels and then restart them with a new connection; also
we are going to be able to easily monitor client accessing channels, queue an
dequeue channels from the pool in order to make them accessible by 1 client at
a time making them race condition free.

[change] move channels to the connection worker
[change] link channels to the connection worker
[add] checkout/checkin channels from a pool making them race condition free
[add] monitor clients to return channels to pool when the client crashes
[add] improve and add some type specs