Keep the RabbitMQ module in the state instead of fetching from env
During the initialization of bugs bunny I split out the amqp related
configs from ours; I pull out the adapter (previously called "client",
which is a loaded term) and store it in the state. There is no need to
pull this from the environment and it is a lot cleaner as pulling it
out of the state is a simple pattern match.